labeling_generation/requirements/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile --output-file='.\requirements.txt' '.\requirements.in'
#
accelerate==1.2.1
    # via -r .\requirements.in
boto3==1.35.93
    # via -r .\requirements.in
botocore==1.35.99
    # via
    #   boto3
    #   s3transfer
certifi==2024.12.14
    # via requests
charset-normalizer==3.4.1
    # via requests
colorama==0.4.6
    # via
    #   loguru
    #   tqdm
filelock==3.16.1
    # via
    #   huggingface-hub
    #   torch
    #   transformers
fsspec==2024.12.0
    # via
    #   huggingface-hub
    #   torch
huggingface-hub==0.27.1
    # via
    #   accelerate
    #   tokenizers
    #   transformers
idna==3.10
    # via requests
jinja2==3.1.5
    # via torch
jmespath==1.0.1
    # via
    #   boto3
    #   botocore
loguru==0.7.3
    # via -r .\requirements.in
markupsafe==3.0.2
    # via jinja2
more-itertools==10.5.0
    # via -r .\requirements.in
mpmath==1.3.0
    # via sympy
networkx==3.4.2
    # via torch
numpy==2.2.1
    # via
    #   accelerate
    #   pandas
    #   transformers
packaging==24.2
    # via
    #   accelerate
    #   huggingface-hub
    #   transformers
pandas==2.2.3
    # via -r .\requirements.in
pillow==11.0.0
    # via -r .\requirements.in
psutil==6.1.1
    # via accelerate
python-dateutil==2.9.0.post0
    # via
    #   botocore
    #   pandas
python-dotenv==1.0.1
    # via -r .\requirements.in
pytz==2024.2
    # via pandas
pyyaml==6.0.2
    # via
    #   accelerate
    #   huggingface-hub
    #   transformers
regex==2024.11.6
    # via transformers
requests==2.32.3
    # via
    #   huggingface-hub
    #   transformers
s3transfer==0.10.4
    # via boto3
safetensors==0.5.2
    # via
    #   accelerate
    #   transformers
six==1.17.0
    # via python-dateutil
sympy==1.13.1
    # via torch
tokenizers==0.21.0
    # via transformers
torch==2.5.1
    # via
    #   -r .\requirements.in
    #   accelerate
tqdm==4.67.1
    # via
    #   -r .\requirements.in
    #   huggingface-hub
    #   transformers
transformers==4.47.0
    # via -r .\requirements.in
typing-extensions==4.12.2
    # via
    #   huggingface-hub
    #   torch
tzdata==2024.2
    # via pandas
urllib3==2.3.0
    # via
    #   botocore
    #   requests
win32-setctime==1.2.0
    # via loguru

# The following packages are considered to be unsafe in a requirements file:
# setuptools
